tests/wpt/tests/css/css-text/parsing/hyphenate-limit-chars-invalid.html

<!DOCTYPE html>
<title>CSS Text: parsing hyphenate-limit-chars with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-text-4/#propdef-hyphenate-limit-chars">
<meta name="assert" content="hyphenate-character supports only the grammar '[ auto | &lt;integer&gt; ]{1,3}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
test_invalid_value("hyphenate-limit-chars", "normal");
test_invalid_value("hyphenate-limit-chars", "1 normal");
test_invalid_value("hyphenate-limit-chars", "none");
test_invalid_value("hyphenate-limit-chars", "1 none");
test_invalid_value("hyphenate-limit-chars", "1.2");
test_invalid_value("hyphenate-limit-chars", "1 1.2");
test_invalid_value("hyphenate-limit-chars", "1 2 1.2");
test_invalid_value("hyphenate-limit-chars", "1 2 3 4");
</script>
